requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
langchain-community
langchain-openai
langchain-core
chromadb
sentence-transformers
apscheduler
PyGithub
beautifulsoup4
requests
pypdf
python-dotenv
sse-starlette
slowapi
